+++ app/scripts/lib/lforms-util.js
@@ -332,20 +332,16 @@ LForms.Util = {
   /**
    *  For FHIR applications, provides FHIR context information that might be
    *  needed in rendering a Quesitonnaire.
+   *  Priort to calling this, the LHC-Forms FHIR support files should be loaded.
    * @param fhirContext an optional object for accessing a FHIR context and
-   *  FHIR API.  It should define the following operations:
-   *  - getCurrent(typeList, callback):  "typeList" should be a list of desired
-   *    FHIR resource types for which there is conceptually a "current" on in
-   *    the FHIR context (e.g., Patient, or Practitioner).  Only one resource
-   *    from the requested list will be returned, and the result will be null if
-   *    none of the requested resource types are available.  Because retrieving
-   *    the resource will generally be an asynchronous operation, the resource
-   *    will be returned via the first argument to the provided "callback"
-   *    function.
-   *  - getFHIRAPI():  Should return an instance of fhir.js for interacting with
-   *    the FHIR server.
+   *  a FHIR API.  It should be an instance of 'client-js', a.k.a. npm package fhirclient,
+   *  version 2.  (See http://docs.smarthealthit.org/client-js).
    */
   setFHIRContext: function(fhirContext) {
+    if (!LForms.FHIR) {
+      throw new Error('LHC-Forms FHIR support files have not been loaded.' +
+        'See http://lhncbc.github.io/lforms/#fhirScripts');
+    }
     LForms.fhirContext = fhirContext;
     LForms.fhirCapabilities = {}; // our own flags about what the server can do
     delete LForms._serverFHIRReleaseID; // in case the version changed
@@ -391,9 +387,9 @@ LForms.Util = {
     if (!LForms._serverFHIRReleaseID) {
       // Retrieve the fhir version
       try {
-        var fhirAPI = LForms.fhirContext.getFHIRAPI();
-        fhirAPI.conformance({}).then(function(res) {
-          var fhirVersion = res.data.fhirVersion;
+        var fhirAPI = LForms.fhirContext;
+        //fhirAPI.request('metadata?_elements=fhirVersion').then(function(res) // causes an error on lforms-smart-fhir (TBD)
+        fhirAPI.getFhirVersion().then(function(fhirVersion) {
           LForms._serverFHIRReleaseID = LForms.Util._fhirVersionToRelease(fhirVersion);
           console.log('Server FHIR version is '+LForms._serverFHIRReleaseID+' ('+
             fhirVersion+')');
@@ -1072,9 +1068,9 @@ LForms.Util = {
 
 
   /**
-   * Get a list of warning messages about answer lists, which should have been 
+   * Get a list of warning messages about answer lists, which should have been
    * loaded from the URL in answerValueSet but were not.
-   * 
+   *
    * @param {*} formDataSource Optional.  Either the containing HTML element that
    *  includes the LForm's rendered form, a CSS selector for that element, an
    *  LFormsData object, or an LForms form definition (parsed).  If not